+++ lib/src/controllers/drawables/drawables.dart
@@ -5,3 +5,4 @@ export 'object_drawable.dart';
 export 'background/background_drawables.dart';
 export 'shape/shape_drawables.dart';
 export 'image_drawable.dart';
+export 'object_group_drawable.dart';